order the code. replace sendMessageContainer and InputMessage, WhriteMessage with SentMessage
