+++ client/src/components/Home.tsx
@@ -1,15 +1,11 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import {
-  StyledHome, Explain, FlexContainer, StyledWrap, SendMessageContainer,
+  StyledHome, Explain, StyledWrap,
 } from './styles/Home.styled';
 import NavBar from './HomeNavBar';
-import InputNumber from './InputNumber';
-import WriteMessage from './WriteMessage';
-import SendButton from './SendButton';
-import ErrorMessage from './styles/ErrorMessage.styled';
+import SendMessage from './SendMessage';
 
 const Home: React.FC = () => {
-  const [showErr, setShowErr] = useState(0);
   useEffect(() => {
     window.addEventListener('load', () => { // make the height static, useful with mobile keyboard
       const viewport = (document.querySelector('meta[name=viewport]') as HTMLMetaElement);
@@ -22,17 +18,7 @@ const Home: React.FC = () => {
       <StyledWrap>
         <NavBar />
         <Explain>שליחת הודעה בוואטספ בלי לשמור את המספר</Explain>
-        <SendMessageContainer>
-          <InputNumber />
-          <ErrorMessage showErr={showErr}>
-            מספר קצר מדי :(
-            מספר תקין הוא משהו בסגנון של 054-123-4567
-          </ErrorMessage>
-          <FlexContainer>
-            <WriteMessage />
-            <SendButton setShowErr={setShowErr} />
-          </FlexContainer>
-        </SendMessageContainer>
+        <SendMessage />
       </StyledWrap>
     </StyledHome>
   );
