--- client/src/components/InputNumber.tsx
@@ -1,39 +1,0 @@
-import { ChangeEvent } from 'react';
-import { useNumberContext } from '../NumberContext';
-import StyledInputNumber from './styles/InputNumber.styled.js';
-import StyledInput from './styles/TextField.styled.js';
-
-const InputNumber: React.FC = () => {
-  const {
-    prefix, phone, changePrefix, changePhone,
-  } = useNumberContext();
-
-  return (
-    <StyledInputNumber>
-      <StyledInput
-        onChange={(e) => changePrefix(e.target.value)}
-        onInput={(e: ChangeEvent<HTMLInputElement>) => {
-          e.target.value = e.target.value.slice(0, 3);
-        }}
-        placeholder="קידומת"
-        type="number"
-        value={prefix}
-        width="25%"
-        ltr={+true}
-      />
-      <StyledInput
-        onChange={(e) => changePhone(e.target.value)}
-        onInput={(e: ChangeEvent<HTMLInputElement>) => {
-          e.target.value = e.target.value.slice(0, 10);
-        }}
-        placeholder="מספר טלפון"
-        type="number"
-        value={phone}
-        width="70%"
-        ltr={+true}
-      />
-    </StyledInputNumber>
-  );
-};
-
-export default InputNumber;
