+++ client/src/components/styles/SendMessage.styled.ts
@@ -2,14 +2,16 @@ import styled from 'styled-components';
 import StyledButton from './Button.styled';
 
 export const StyledSendMessage = styled.div`
-height: 25vh;
+height: 33vh;
 display: flex;
 justify-content: space-between;
+margin-bottom: 20px;
 `;
 
 export const InputContainer = styled.div`
 width: 70%;
 display: grid;
+align-items: flex-end;
 grid-template-columns: 3.2fr 1fr;
 grid-template-rows: 1fr 1fr;
 grid-gap: 10% 5%;
@@ -17,6 +19,7 @@ grid-gap: 10% 5%;
 
 export const SendButton = styled(StyledButton)`
 && {
+  align-self: flex-end;
   width: 20%;
   font-size: 0.9rem;
 }
